feat: Completed the exercise for node #6
Completed the exercise for the sixth node, lesson #4. This
involved signing a release build of an app. In order to do
this, a keystore was created with a single key and build files
were adjusted for the release build to use the keystore and key.
Finally, a keystore.properties file was used to store all of the
settings within the keystore and key and these properties were
used within the build script. The .gitignore was also adjusted
for the keystore properties file to not be included in git
commits.